// src/type/pull-request.type.ts
export type ReviewStatus =
  | 'PENDING_REVIEW'
  | 'IN_REVIEW'
  | 'CHANGES_REQUESTED'
  | 'APPROVED'
  | 'NEEDS_WORK'
  | 'DRAFT';

export interface IReviewFeedback {
  section?: string;
  rating?: number; // 1-5
  comment?: string;
}